+++ rest-api-server/src/integration/kotlin/eu/tib/orkg/prototype/community/application/ObservatoryControllerTest.kt
@@ -11,7 +11,7 @@ import eu.tib.orkg.prototype.createResource
 import eu.tib.orkg.prototype.createUser
 import eu.tib.orkg.prototype.statements.api.ClassUseCases
 import eu.tib.orkg.prototype.statements.api.ResourceUseCases
-import eu.tib.orkg.prototype.statements.application.ResourceControllerTest
+import eu.tib.orkg.prototype.statements.application.ResourceControllerIntegrationTest
 import eu.tib.orkg.prototype.statements.application.RestDocumentationBaseTest
 import eu.tib.orkg.prototype.statements.auth.MockUserDetailsService
 import org.assertj.core.api.Assertions.*
@@ -124,7 +124,7 @@ class ObservatoryControllerTest : RestDocumentationBaseTest() {
             .andDo(
                 document(
                     snippet,
-                    ResourceControllerTest.listOfResourcesResponseFields()
+                    ResourceControllerIntegrationTest.listOfResourcesResponseFields()
                 )
             )
     }
@@ -150,7 +150,7 @@ class ObservatoryControllerTest : RestDocumentationBaseTest() {
             .andDo(
                 document(
                     snippet,
-                    ResourceControllerTest.listOfResourcesResponseFields()
+                    ResourceControllerIntegrationTest.listOfResourcesResponseFields()
                 )
             )
     }
@@ -176,7 +176,7 @@ class ObservatoryControllerTest : RestDocumentationBaseTest() {
             .andDo(
                 document(
                     snippet,
-                    ResourceControllerTest.listOfResourcesResponseFields()
+                    ResourceControllerIntegrationTest.listOfResourcesResponseFields()
                 )
             )
     }