+++ rest-api-server/src/integration/kotlin/eu/tib/orkg/prototype/statements/application/StatementControllerTest.kt
@@ -6,7 +6,7 @@ import eu.tib.orkg.prototype.statements.api.ResourceUseCases
 import eu.tib.orkg.prototype.statements.api.StatementUseCases
 import eu.tib.orkg.prototype.statements.application.LiteralControllerTest.RestDoc.literalResponseFields
 import eu.tib.orkg.prototype.statements.application.PredicateControllerTest.RestDoc.predicateResponseFields
-import eu.tib.orkg.prototype.statements.application.ResourceControllerTest.RestDoc.resourceResponseFields
+import eu.tib.orkg.prototype.statements.application.ResourceControllerIntegrationTest.RestDoc.resourceResponseFields
 import eu.tib.orkg.prototype.statements.auth.MockUserDetailsService
 import org.assertj.core.api.Assertions.assertThat
 import org.hamcrest.Matchers.hasSize